Adjust batting average,score,game end comment
* Member batting average
- Modified to generate a pitching result through the batting average
  of each member.

* score
- Modified to increase 1 point after 4 hits in a round.
- Modified to save score about each team.

* end comment
- Modified to show game end comment (게임 종료)
- Modified to show team name.
- Modified to show team score.

+++ baseballgamemanager.cpp
@@ -139,7 +139,7 @@ void BaseballGameManager::startGame()
     m_pStatusPrinter->showGameEndComment(m_pHomeTeam->getName(), m_pAwayTeam->getName(), m_pScoreBoard->getScore(true), m_pScoreBoard->getScore(false));
 }
 
-bool BaseballGameManager::playAttack(float battingAverage)
+bool BaseballGameManager::playAttack(double battingAverage)
 {
     bool bEndTheAtBat = false;
